GAL_001_002 moved Permission Tests to its own Test Class [SR, AE]
